--- a/spec/spec_helper.rb
+++ b/spec/spec_helper.rb
@@ -36,6 +36,113 @@ module RspecHelper
   end
 end
 
+module LramaCustomMatchers
+  class SymbolMatcher
+    attr_reader :expected, :target
+
+    def initialize(expected)
+      @expected = expected
+      @_failure_message = nil
+    end
+
+    def matches?(target)
+      @target = target
+
+      if !@expected.is_a?(Lrama::Grammar::Symbol)
+        @_failure_message = "expected #{@expected.inspect} to be Lrama::Grammar::Symbol"
+        return false
+      end
+
+      if !@target.is_a?(Lrama::Grammar::Symbol)
+        @_failure_message = "expected #{@target.inspect} to be Lrama::Grammar::Symbol"
+        return false
+      end
+
+      @expected.id == @target.id &&
+      @expected.alias_name == @target.alias_name &&
+      @expected.number == @target.number &&
+      @expected.tag == @target.tag &&
+      @expected.term == @target.term &&
+      @expected.token_id == @target.token_id &&
+      @expected.nullable == @target.nullable &&
+      @expected.precedence == @target.precedence &&
+      @expected.printer == @target.printer &&
+      @expected.error_token == @target.error_token
+    end
+
+    def failure_message
+      return @_failure_message if @_failure_message
+
+      "expected #{@target.inspect} to match with #{@expected.inspect}"
+    end
+
+    def failure_message_when_negated
+      return @_failure_message if @_failure_message
+
+      "expected #{@target.inspect} not to match with #{@expected.inspect}"
+    end
+  end
+
+  class SymbolsMatcher
+    attr_reader :expected, :target
+
+    def initialize(expected)
+      @expected = expected
+      @_failure_message = nil
+    end
+
+    def matches?(target)
+      @target = target
+
+      if !@expected.is_a?(Array)
+        @_failure_message = "expected #{@expected.inspect} to be Array"
+        return false
+      end
+
+      if !@target.is_a?(Array)
+        @_failure_message = "expected #{@target.inspect} to be Array"
+        return false
+      end
+
+      if @expected.count != @target.count
+        @_failure_message = "expected the number of array to be same (#{@expected.count} != #{@target.count})"
+        return false
+      end
+
+      @not_matched = []
+
+      @expected.zip(@target).each do |expected, actual|
+        matcher = SymbolMatcher.new(expected)
+        unless matcher.matches?(actual)
+          @not_matched << matcher
+        end
+      end
+
+      @not_matched.empty?
+    end
+
+    def failure_message
+      return @_failure_message if @_failure_message
+
+      @not_matched.map(&:failure_message).join("\n")
+    end
+
+    def failure_message_when_negated
+      return @_failure_message if @_failure_message
+
+      @not_matched.map(&:failure_message_when_negated).join("\n")
+    end
+  end
+
+  def match_symbol(expected)
+    SymbolMatcher.new(expected)
+  end
+
+  def match_symbols(expected)
+    SymbolsMatcher.new(expected)
+  end
+end
+
 RSpec.configure do |config|
   # Disable RSpec exposing methods globally on `Module` and `main`
   config.disable_monkey_patching!
@@ -45,6 +152,7 @@ RSpec.configure do |config|
   end
 
   config.include(RspecHelper)
+  config.include(LramaCustomMatchers)
 
   # Allow to limit the run of the specs
   # NOTE: Please do not commit the filter option.